+++ CHANGELOG.md
@@ -1,10 +1,21 @@
+## [1.2.0] - 2020/06/25
+
+* Breaking Change: ColumnSuper width is now the max intrinsic width of its children, just like a regular Column. 
+  To restore old behavior: `Container(width: double.infinity, child: ColumnSuper(...))`.
+* Breaking Change: RowSuper height is now the max intrinsic height of its children, just like a regular Row. 
+  To restore old behavior: `Container(height: double.infinity, child: RowSuper(...))`.
+* New examples: 
+<a href="https://github.com/marcglasberg/assorted_layout_widgets/blob/master/example/lib/main_column_super_playground.dart">ColumnSuper Playground</a> 
+and 
+<a href="https://github.com/marcglasberg/assorted_layout_widgets/blob/master/example/lib/main_row_super_playground.dart">RowSuper Playground</a>.   
+
 ## [1.1.4] - 2020/06/24
 
 * Fix: ColumnSuper intrinsic height, and RowSuper intrinsic width.
 
 ## [1.1.3] - 2020/06/15
 
-* Fix: WrapSuper minimum raggedness algorithm now uses the correct JavaScript's Number.MAX_SAFE_INTEGER.
+* Fix: WrapSuper minimum raggedness algorithm now uses the correct JavaScript's `Number.MAX_SAFE_INTEGER`.
 * Fix: Divide by zero conditions.
 
 ## [1.1.1] - 2020/05/19